feat(layer load): move code for layer load to a specific class LayerLoad
