+++ menu_from_project/menu_from_project.py
@@ -21,11 +21,10 @@ email                : xavier.culos@eau-adour-garonne.fr
 # Standard library
 import logging
 import os
-import re
-from pathlib import Path
-from typing import Any, Dict, List, Optional, Tuple
+from typing import Dict, Optional, List, Tuple, Any
 
 # PyQGIS
+from menu_from_project.logic.layer_load import LayerLoad
 from menu_from_project.toolbelt.preferences import (
     SOURCE_MD_LAYER,
     SOURCE_MD_NOTE,
@@ -35,42 +34,32 @@ from menu_from_project.toolbelt.preferences import (
 from qgis.core import (
     QgsApplication,
     QgsMessageLog,
-    QgsProject,
-    QgsRasterLayer,
-    QgsReadWriteContext,
     QgsSettings,
-    QgsVectorLayer,
-    QgsVectorTileLayer,
-    QgsRelation,
     QgsTask,
 )
-from qgis.PyQt.QtCore import QCoreApplication, QFileInfo, Qt, QTranslator, QUuid
+from qgis.PyQt.QtCore import QCoreApplication, QFileInfo, Qt, QTranslator
 from qgis.PyQt.QtGui import QFont, QIcon
-from qgis.PyQt.QtWidgets import QAction, QMenu, QWidget
+from qgis.PyQt.QtWidgets import QAction, QMenu
 from qgis.PyQt.QtCore import QLocale, QUrl, QDir
 from qgis.PyQt.QtGui import QDesktopServices
-from qgis.utils import plugins
 
 # project
-from .__about__ import DIR_PLUGIN_ROOT, __title__, __title_clean__
-from .logic.qgs_manager import (
+from menu_from_project.__about__ import DIR_PLUGIN_ROOT, __title__, __title_clean__
+from menu_from_project.logic.qgs_manager import (
     QgsDomManager,
-    is_absolute,
     read_from_file,
     read_from_http,
 )
-from .logic.tools import (
-    guess_type_from_uri,
+from menu_from_project.logic.tools import (
     icon_per_layer_type,
 )
-from .ui.menu_conf_dlg import MenuConfDialog  # noqa: F4 I001
+from menu_from_project.ui.menu_conf_dlg import MenuConfDialog  # noqa: F4 I001
 from menu_from_project.logic.project_read import (
     MenuGroupConfig,
     MenuLayerConfig,
     MenuProjectConfig,
     get_project_menu_config,
 )
-from menu_from_project.logic.xml_utils import getFirstChildByTagNameValue
 
 # ############################################################################
 # ########## Globals ###############
@@ -117,24 +106,6 @@ def showPluginHelp(packageName: str = None, filename: str = "index", section: st
         if section != "":
             url = url + "#" + section
         QDesktopServices.openUrl(QUrl(url, QUrl.TolerantMode))
-
-
-def project_trusted(doc):
-    """Return if the project is trusted.
-
-    :param doc: The QGIS project as XML document. Default to None.
-    :type doc: QDomDocument
-
-    :return: True of False.
-    :rtype: bool
-    """
-    tags = doc.elementsByTagName("qgis")
-    if tags.count():
-        node = tags.at(0)
-        trust_node = node.namedItem("trust")
-        return trust_node.toElement().attribute("active") == "1"
-
-    return False
 
 
 # ############################################################################
@@ -376,7 +347,7 @@ class MenuFromProject:
                 action.setFont(font)
                 grp_menu.addAction(action)
                 action.triggered.connect(
-                    lambda checked, f=None, w=None, m=grp_menu: self.loadLayer(
+                    lambda checked, f=None, w=None, m=grp_menu: LayerLoad().loadLayer(
                         None, f, w, m
                     )
                 )
@@ -396,7 +367,7 @@ class MenuFromProject:
 
         # add menu item
         action.triggered.connect(
-            lambda checked, uri=layer.filename, f=layer.filename, lid=layer.layer_id, m=menu, v=layer.visible, x=layer.expanded: self.loadLayer(
+            lambda checked, uri=layer.filename, f=layer.filename, lid=layer.layer_id, m=menu, v=layer.visible, x=layer.expanded: LayerLoad().loadLayer(
                 uri, f, lid, m, v, x
             )
         )
@@ -511,369 +482,3 @@ class MenuFromProject:
 
             # build menus
             self.initMenus()
-
-    def addLayer(
-        self,
-        uri,
-        doc,
-        layerId,
-        group=None,
-        visible=False,
-        expanded=False,
-        parentsLoop: dict = {},
-        loop=0,
-    ):
-        theLayer = None
-        settings = self.plg_settings.get_plg_settings()
-
-        # is project in relative path ?
-        absolute = is_absolute(doc)
-        trusted = project_trusted(doc)
-
-        node = getFirstChildByTagNameValue(
-            doc.documentElement(), "maplayer", "id", layerId
-        )
-        node = node.cloneNode()
-        if node:
-            idNode = node.namedItem("id")
-            layerType = node.toElement().attribute("type", "vector")
-            # give it a new id (for multiple import)
-            newLayerId = "L%s" % re.sub("[{}-]", "", QUuid.createUuid().toString())
-            self.mapLayerIds[newLayerId] = layerId
-
-            try:
-                idNode.firstChild().toText().setData(newLayerId)
-            except Exception:
-                pass
-
-            # if relative path, adapt datasource
-            if not absolute:
-                try:
-                    datasourceNode = node.namedItem("datasource")
-                    ds = datasourceNode.firstChild().toText().data()
-                    providerNode = node.namedItem("provider")
-                    provider = providerNode.firstChild().toText().data()
-
-                    if provider in ["ogr", "gdal"] and (ds.find(".") == 0):
-                        projectpath = QFileInfo(uri).path()
-                        newlayerpath = projectpath + "/" + ds
-                        datasourceNode.firstChild().toText().setData(newlayerpath)
-                except Exception:
-                    pass
-
-            # is relations exists ?
-            relationsToBuild = []
-            if settings.optionOpenLinks:
-                relationsToBuild = self.buildRelations(
-                    uri, doc, layerId, newLayerId, group, parentsLoop, loop
-                )
-
-            # read modified layer node
-            newLayer = None
-            if settings.optionCreateGroup and group is not None:
-                if layerType == "raster":
-                    theLayer = QgsRasterLayer()
-                elif layerType == "vector-tile":
-                    theLayer = QgsVectorTileLayer()
-                else:
-                    theLayer = QgsVectorLayer()
-                    theLayer.setReadExtentFromXml(trusted)
-
-                theLayer.readLayerXml(node.toElement(), QgsReadWriteContext())
-
-                # Special process if the plugin "DB Style Manager" is installed
-                flag = "use_db_style_manager_in_custom_menu" in os.environ
-                if flag and "db-style-manager" in plugins:
-                    try:
-                        plugins["db-style-manager"].load_style_from_database(theLayer)
-                    except Exception:
-                        self.log(
-                            "DB-Style-Manager failed to load the style.",
-                            indent=loop,
-                        )
-
-                # needed
-                newLayer = QgsProject.instance().addMapLayer(theLayer, False)
-                if newLayer is not None:
-                    # add to group
-                    treeNode = group.insertLayer(0, newLayer)
-                    treeNode.setExpanded(expanded)
-                    treeNode.setItemVisibilityChecked(visible)
-            else:
-                # create layer
-                ok = QgsProject.instance().readLayer(node)
-                if ok:
-                    newLayer = QgsProject.instance().mapLayer(newLayerId)
-
-            return newLayer, relationsToBuild
-
-        else:
-            self.log("{} not found".format(layerId), indent=loop)
-
-        return None, None
-
-    def getRelations(self, doc):
-        """
-        Charger la définition des relations (niveau projet), pour les rétablir éventuellement après chargement des couches
-
-        <relations>
-            <relation strength="Association" referencedLayer="layerid" id="refid" name="fk_region" referencingLayer="layerid">
-                <fieldRef referencedField="insee_region" referencingField="insee_region"/>
-            </relation>
-        </relations>
-        """
-        relations = []
-        try:
-            nodes = doc.elementsByTagName("relations")
-            relsNode = nodes.at(0)
-
-            relNodes = relsNode.toElement().elementsByTagName("relation")
-            for relNode in (relNodes.at(i) for i in range(relNodes.size())):
-                fieldNodes = relNode.toElement().elementsByTagName("fieldRef")
-                fieldNode = fieldNodes.at(0)
-
-                if fieldNode:
-                    relation = {}
-                    for attr in [
-                        "strength",
-                        "referencedLayer",
-                        "id",
-                        "name",
-                        "referencingLayer",
-                    ]:
-                        relation[attr] = relNode.toElement().attribute(attr)
-
-                    for attr in [
-                        "referencedField",
-                        "referencingField",
-                    ]:
-                        relation[attr] = fieldNode.toElement().attribute(attr)
-
-                    if relation["referencedLayer"] != "":
-                        relations.append(relation)
-        except Exception as e:
-            for m in e.args:
-                self.log(m)
-
-        return relations
-
-    def getRelationsForLayer(self, relations, source=None, target=None):
-        """Retourne le dico de la relation selon si 'source'=referencedLayer ou 'target'=referencingLayer"""
-
-        r = []
-        try:
-            for relation in relations:
-                if source is not None and source == relation["referencedLayer"]:
-                    r.append(relation)
-
-                if target is not None and target == relation["referencingLayer"]:
-                    r.append(relation)
-        except Exception as e:
-            for m in e.args:
-                self.log(m)
-
-        return r
-
-    def fixForm(self, doc, newLayerId: str, oldRelationId, newRelationId):
-        """rebuilds the form when relations are defined in it
-
-        Principle: reading the source XML document, updating identifiers, updating editFormConfig
-        """
-        theLayer = QgsProject.instance().mapLayer(newLayerId)
-        oldLayerId = self.mapLayerIds[newLayerId]
-
-        layerNode = getFirstChildByTagNameValue(
-            doc.documentElement(), "maplayer", "id", oldLayerId
-        )
-
-        nodes = layerNode.toElement().elementsByTagName("attributeEditorForm")
-        if nodes.count() == 0:
-            return
-        aefNode = nodes.at(0)
-
-        nodes = aefNode.toElement().elementsByTagName("attributeEditorRelation")
-        for nodeIdx in range(nodes.length()):
-            aerNode = nodes.at(nodeIdx)
-            rid = aerNode.toElement().attribute("relation")
-            if rid == oldRelationId:
-                aerNode.toElement().setAttribute("relation", newRelationId)
-
-                nodes = aefNode.toElement().elementsByTagName("widgets")
-                widgets = nodes.at(0)
-                widgets.toElement().setAttribute("name", newRelationId)
-
-                editFormConfig = theLayer.editFormConfig()
-                rootContainer = editFormConfig.invisibleRootContainer()
-                rootContainer.clear()
-                editFormConfig.clearTabs()
-                editFormConfig.readXml(layerNode, QgsReadWriteContext())
-                theLayer.setEditFormConfig(editFormConfig)
-
-    def buildProjectRelation(self, doc, relDict):
-        try:
-            """builds one relation, add it to the project"""
-            REL_STRENGTH = {
-                "Association": QgsRelation.Association,
-                "Composition": QgsRelation.Composition,
-            }
-            relMan = QgsProject.instance().relationManager()
-
-            rel = QgsRelation()
-            rel.addFieldPair(relDict["referencingField"], relDict["referencedField"])
-            oldRelationId = relDict["id"]
-            newRelationId = "R%s" % re.sub("[{}-]", "", QUuid.createUuid().toString())
-            rel.setId(newRelationId)
-            rel.setName(relDict["name"])
-            rel.setReferencedLayer(relDict["referencedLayer"])
-            rel.setReferencingLayer(relDict["referencingLayer"])
-            rel.setStrength(REL_STRENGTH[relDict["strength"]])
-            rel.updateRelationStatus()
-
-            if rel.isValid():
-                relMan.addRelation(rel)
-
-                # Adapter le formulaire de la couche referencedLayer
-                try:
-                    self.fixForm(
-                        doc,
-                        relDict["referencedLayer"],
-                        oldRelationId,
-                        newRelationId,
-                    )
-                except Exception:
-                    self.log(
-                        "Form not fixed for layer {}".format(relDict["referencedLayer"])
-                    )
-
-            else:
-                self.log(
-                    "Invalid relation {} : {}".format(rel.id(), rel.validationError())
-                )
-        except Exception as e:
-            for m in e.args:
-                self.log(m)
-
-    def buildRelations(
-        self, uri, doc, oldLayerId, newLayerId, group, parentsLoop, loop
-    ):
-        """identify the relations to be created (later, after source layer creation)
-
-        Based on those of the source project, adapted to the new identifiers of the layers
-        """
-        relationsToBuild, targetRelations = [], []
-
-        relations = self.getRelations(doc)
-        relsTarget = self.getRelationsForLayer(relations, source=oldLayerId)
-        # relsSource = self.getRelationsForLayer(relations, target=oldLayerId)
-
-        if len(relsTarget) > 0:
-            for relDict in relsTarget:
-                if relDict["referencingLayer"] in parentsLoop:
-                    # La couche cible a déjà été ajoutée (boucle infinie)
-                    # on se contente de référencer celle-ci
-                    relDict["referencedLayer"] = newLayerId
-                    relDict["referencingLayer"] = parentsLoop[
-                        relDict["referencingLayer"]
-                    ]
-                    relationsToBuild.append(relDict)
-                else:
-                    # la couche cible n'a pas été ajoutée
-                    parentsLoop.update({oldLayerId: newLayerId})
-
-                    targetLayer, targetRelations = self.addLayer(
-                        uri,
-                        doc,
-                        relDict["referencingLayer"],
-                        group,
-                        False,
-                        False,
-                        parentsLoop,
-                        loop + 1,
-                    )
-                    if targetLayer is not None:
-                        relDict["referencedLayer"] = newLayerId
-                        relDict["referencingLayer"] = targetLayer.id()
-                        relationsToBuild.append(relDict)
-
-        return targetRelations + relationsToBuild
-
-    def loadLayer(self, uri, fileName, layerId, menu=None, visible=None, expanded=None):
-        """Load the chosen layer(s)
-
-        :param uri: The layer URI (file path or PG URI)
-        :type uri: basestring
-
-        :param layerId: The layer ID to look for in the project.
-        :type layerId: basestring
-
-        """
-        self.canvas.freeze(True)
-        self.canvas.setRenderFlag(False)
-        group = None
-        QgsApplication.setOverrideCursor(Qt.WaitCursor)
-        self.mapLayerIds = {}
-
-        settings = self.plg_settings.get_plg_settings()
-
-        try:
-            if (
-                isinstance(menu.parentWidget(), (QMenu, QWidget))
-                and settings.optionCreateGroup
-            ):
-                groupName = menu.title().replace("&", "")
-                group = QgsProject.instance().layerTreeRoot().findGroup(groupName)
-                if group is None:
-                    group = (
-                        QgsProject.instance().layerTreeRoot().insertGroup(0, groupName)
-                    )
-
-            # load all layers
-            if fileName is None and layerId is None and settings.optionLoadAll:
-                for action in menu.actions()[::-1]:
-                    if (
-                        action.text() != self.tr("Load all")
-                        and action.text() != "Load all"
-                    ):
-                        action.trigger()
-            else:
-                doc, _ = self.qgs_dom_manager.getQgsDoc(fileName)
-
-                # Loading layer
-                layer, relationsToBuild = self.addLayer(
-                    uri, doc, layerId, group, visible, expanded, {}, 0
-                )
-                for relDict in relationsToBuild:
-                    self.buildProjectRelation(doc, relDict)
-
-                # is joined layers exists ?
-                if settings.optionOpenLinks and layer and type(layer) == QgsVectorLayer:
-                    for j in layer.vectorJoins():
-                        try:
-                            joinLayer, joinRelations = self.addLayer(
-                                uri, doc, j.joinLayerId(), group
-                            )
-                            for relDict in joinRelations:
-                                self.buildProjectRelation(doc, relDict)
-
-                            if joinLayer:
-                                j.setJoinLayerId(joinLayer.id())
-                                j.setJoinLayer(joinLayer)
-                                layer.addJoin(j)
-                        except Exception as e:
-                            self.log(
-                                "Joined layer {} not added.".format(j.joinLayerId())
-                            )
-                            pass
-
-        except Exception as e:
-            # fixme fileName is not defined
-            # self.log(
-            #     'Menu from layer: Invalid ' + (fileName if fileName is not None else ""))
-            for m in e.args:
-                self.log(m)
-
-        finally:
-            self.canvas.freeze(False)
-            self.canvas.setRenderFlag(True)
-            self.canvas.refresh()
-            QgsApplication.restoreOverrideCursor()
